Updated host_template and gateway provisioning scripts to work with clusters.

+++ lxd/gateway/destroy_lxc_gateway.sh
@@ -5,16 +5,33 @@ cd "$(dirname "$0")"
 source ./defaults.sh
 
 
-
+MASTER_NODE=$(lxc info | grep server_name | xargs | awk 'NF>1{print $NF}')
 for endpoint in $(bcm cluster list --endpoints --cluster-name=$BCM_CLUSTER_NAME); do
     #echo $endpoint
     HOST_ENDING=$(echo $endpoint | tail -c 2)
-    LXD_CONTAINER_NAME=bcm-gateway-$(printf %02d $HOST_ENDING)
+    LXD_CONTAINER_NAME="bcm-gateway-$(printf %02d $HOST_ENDING)"
+    #echo "HOST_ENDING: $HOST_ENDING"
+    #echo "LXD_CONTAINER_NAME: $LXD_CONTAINER_NAME"
     # let's kill a bcm-gateway LXC instance on each cluster endpoint.
-    if [[ ! -z $(lxc list | grep $LXD_CONTAINER_NAME) ]]; then
-        lxc delete $LXD_CONTAINER_NAME --force
+    if [[ $MASTER_NODE != $endpoint ]]; then
+        if [[ ! -z $(lxc list | grep "$endpoint") ]]; then
+            lxc delete $LXD_CONTAINER_NAME --force
+        fi
+
+        if [[ ! -z $(lxc storage volume list bcm_btrfs | grep "bcm-gateway-dockervol") ]]; then
+            lxc storage volume delete bcm_btrfs bcm-gateway-dockervol --target $endpoint
+        fi
     fi
 done
+
+MASTER_NODE="bcm-gateway-01"
+if [[ ! -z $(lxc list | grep $MASTER_NODE) ]]; then
+    lxc delete $MASTER_NODE --force
+fi
+
+if [[ ! -z $(lxc storage volume list bcm_btrfs | grep "bcm-gateway-dockervol") ]]; then
+    lxc storage volume delete bcm_btrfs bcm-gateway-dockervol
+fi
 
 if [[ $BCM_REMOVE_TEMPLATE_FLAG == 1 ]]; then
     # delte container gateway-template
@@ -34,10 +51,10 @@ if [[ $BCM_REMOVE_TEMPLATE_FLAG == 1 ]]; then
 fi
 
 
-
-# this deletes the image from the cluster.
-lxc image delete bcm-gateway-template
-
+if [[ ! -z $(lxc image list | grep "bcm-gateway-template") ]]; then
+    # this deletes the image from the cluster.
+    lxc image delete bcm-gateway-template
+fi
 
 # delete the profile bcm-gateway-profile
 if [[ $BCM_GATEWAY_PROFILE_GATEWAYPROFILE_DELETE = 1 ]]; then
